Solved an issue with previous landing controller name (renamed to LandingPage)
